OCPBUGS-44576: Resolve min-length cycle in `cluster-settings`
