+++ frontend/public/components/cluster-settings/cluster-settings.tsx
@@ -19,13 +19,11 @@ import {
   TextVariants,
 } from '@patternfly/react-core';
 import { Link } from 'react-router-dom-v5-compat';
-import { HashLink } from 'react-router-hash-link';
 import { useTranslation } from 'react-i18next';
 
 import { AddCircleOIcon } from '@patternfly/react-icons/dist/esm/icons/add-circle-o-icon';
 import { PauseCircleIcon } from '@patternfly/react-icons/dist/esm/icons/pause-circle-icon';
 import { PencilAltIcon } from '@patternfly/react-icons/dist/esm/icons/pencil-alt-icon';
-import { SyncAltIcon } from '@patternfly/react-icons/dist/esm/icons/sync-alt-icon';
 
 import { removeQueryArgument } from '@console/internal/components/utils/router';
 import { SyncMarkdownView } from '@console/internal/components/markdown-view';
@@ -55,13 +53,11 @@ import {
   clusterIsUpToDateOrUpdateAvailable,
   ClusterOperator,
   ClusterUpdateStatus,
-  ClusterVersionConditionType,
   ClusterVersionKind,
   clusterVersionReference,
   getClusterID,
   getClusterOperatorVersion,
   getClusterUpdateStatus,
-  getClusterVersionCondition,
   getConditionUpgradeableFalse,
   getCurrentVersion,
   getDesiredClusterVersion,
@@ -78,7 +74,6 @@ import {
   isMCPPaused,
   isMCPWorker,
   isMinorVersionNewer,
-  k8sPatch,
   K8sResourceConditionStatus,
   K8sResourceKind,
   MachineConfigPoolConditionType,
@@ -108,17 +103,12 @@ import {
   SectionHeading,
   Timestamp,
   togglePaused,
-  truncateMiddle,
   UpstreamConfigDetailsItem,
   useAccessReview,
 } from '../utils';
 import { useK8sWatchResource } from '@console/internal/components/utils/k8s-watch-hook';
 import {
-  BlueArrowCircleUpIcon,
-  BlueInfoCircleIcon,
-  GreenCheckCircleIcon,
   isClusterExternallyManaged,
-  RedExclamationCircleIcon,
   useCanClusterUpgrade,
   YellowExclamationTriangleIcon,
 } from '@console/shared';
@@ -133,15 +123,7 @@ import {
   ServiceLevelLoading,
 } from '../utils/service-level';
 import { hasAvailableUpdates, hasNotRecommendedUpdates } from '../../module/k8s/cluster-settings';
-
-const cancelUpdate = (cv: ClusterVersionKind) => {
-  k8sPatch(ClusterVersionModel, cv, [{ path: '/spec/desiredUpdate', op: 'remove' }]).catch(
-    (err) => {
-      const error = err.message;
-      errorModal({ error });
-    },
-  );
-};
+import { UpdateStatus } from './cluster-settings-messages';
 
 export const clusterAutoscalerReference = referenceForModel(ClusterAutoscalerModel);
 
@@ -204,170 +186,6 @@ export const CurrentChannel: React.FC<CurrentChannelProps> = ({ cv, canUpgrade }
   ) : (
     <>{label}</>
   );
-};
-
-const StatusMessagePopover: React.FC<CVStatusMessagePopoverProps> = ({ bodyContent, children }) => {
-  return (
-    <Popover bodyContent={truncateMiddle(bodyContent, { length: 256 })}>
-      <Button variant="link" isInline>
-        <span>{children}</span>
-      </Button>
-    </Popover>
-  );
-};
-
-const InvalidMessage: React.FC<CVStatusMessageProps> = ({ cv }) => {
-  const { t } = useTranslation();
-  return (
-    <div data-test="cv-update-status-invalid">
-      <div>
-        <RedExclamationCircleIcon /> {t('public~Invalid cluster version')}
-      </div>
-      <Button onClick={() => cancelUpdate(cv)} variant="primary" className="pf-v5-u-mt-xs">
-        {t('public~Cancel update')}
-      </Button>
-    </div>
-  );
-};
-
-const ReleaseNotAcceptedMessage: React.FC<CVStatusMessageProps> = ({ cv }) => {
-  const releaseNotAcceptedCondition = getClusterVersionCondition(
-    cv,
-    ClusterVersionConditionType.ReleaseAccepted,
-    K8sResourceConditionStatus.False,
-  );
-  const { t } = useTranslation();
-  return (
-    <>
-      <div data-test="cv-update-status-release-accepted-false">
-        <StatusMessagePopover bodyContent={releaseNotAcceptedCondition.message}>
-          <RedExclamationCircleIcon /> {t('public~Release not accepted')}
-        </StatusMessagePopover>
-      </div>
-      <ClusterVersionConditionsLink cv={cv} />
-    </>
-  );
-};
-
-const UpdatesAvailableMessage: React.FC<CVStatusMessageProps> = () => {
-  const { t } = useTranslation();
-  return (
-    <div className="co-update-status" data-test="cv-update-status-available-updates">
-      <BlueArrowCircleUpIcon /> {t('public~Available updates')}
-    </div>
-  );
-};
-
-const FailingMessageText: React.FC<CVStatusMessageProps> = ({ cv }) => {
-  const failingCondition = getClusterVersionCondition(
-    cv,
-    ClusterVersionConditionType.Failing,
-    K8sResourceConditionStatus.True,
-  );
-  const { t } = useTranslation();
-  return (
-    <div data-test="cv-update-status-failing">
-      <StatusMessagePopover bodyContent={failingCondition.message}>
-        <RedExclamationCircleIcon /> {t('public~Failing')}
-      </StatusMessagePopover>
-    </div>
-  );
-};
-
-export const ClusterVersionConditionsLink: React.FC<ClusterVersionConditionsLinkProps> = ({
-  cv,
-}) => {
-  const { t } = useTranslation();
-  return (
-    <HashLink
-      smooth
-      to={`${resourcePathFromModel(ClusterVersionModel, cv.metadata.name)}#conditions`}
-    >
-      {t('public~View conditions')}
-    </HashLink>
-  );
-};
-
-export const UpdatingMessageText: React.FC<CVStatusMessageProps> = ({ cv }) => {
-  const version = getDesiredClusterVersion(cv);
-  const { t } = useTranslation();
-  return <>{t('public~Update to {{version}} in progress', { version })}</>;
-};
-
-const UpdatingMessage: React.FC<CVStatusMessageProps> = ({ cv, isFailing }) => {
-  return (
-    <>
-      <div data-test="cv-update-status-updating">
-        <SyncAltIcon className="fa-spin co-icon-space-r" />
-        <UpdatingMessageText cv={cv} />
-      </div>
-      {isFailing && <FailingMessageText cv={cv} />}
-      <ClusterVersionConditionsLink cv={cv} />
-    </>
-  );
-};
-
-const ErrorRetrievingMessage: React.FC<CVStatusMessageProps> = ({ cv }) => {
-  const retrievedUpdatesCondition = getClusterVersionCondition(
-    cv,
-    ClusterVersionConditionType.RetrievedUpdates,
-    K8sResourceConditionStatus.False,
-  );
-  const { t } = useTranslation();
-  return retrievedUpdatesCondition.reason === 'NoChannel' ? (
-    <div data-test="cv-update-status-no-channel">
-      <BlueInfoCircleIcon /> {retrievedUpdatesCondition.message}
-    </div>
-  ) : (
-    <>
-      <div data-test="cv-update-status-no-updates">
-        <StatusMessagePopover bodyContent={retrievedUpdatesCondition.message}>
-          <RedExclamationCircleIcon /> {t('public~Not retrieving updates')}
-        </StatusMessagePopover>
-      </div>
-      <ClusterVersionConditionsLink cv={cv} />
-    </>
-  );
-};
-
-const FailingMessage: React.FC<CVStatusMessageProps> = ({ cv }) => {
-  return (
-    <>
-      <FailingMessageText cv={cv} />
-      <ClusterVersionConditionsLink cv={cv} />
-    </>
-  );
-};
-
-export const UpToDateMessage: React.FC<{}> = () => {
-  const { t } = useTranslation();
-  return (
-    <span data-test="cv-update-status-up-to-date">
-      <GreenCheckCircleIcon /> {t('public~Up to date')}
-    </span>
-  );
-};
-
-export const UpdateStatus: React.FC<UpdateStatusProps> = ({ cv }) => {
-  const status = getClusterUpdateStatus(cv);
-  switch (status) {
-    case ClusterUpdateStatus.Invalid:
-      return <InvalidMessage cv={cv} />;
-    case ClusterUpdateStatus.ReleaseNotAccepted:
-      return <ReleaseNotAcceptedMessage cv={cv} />;
-    case ClusterUpdateStatus.UpdatesAvailable:
-      return <UpdatesAvailableMessage cv={cv} />;
-    case ClusterUpdateStatus.Updating:
-      return <UpdatingMessage cv={cv} />;
-    case ClusterUpdateStatus.UpdatingAndFailing:
-      return <UpdatingMessage cv={cv} isFailing />;
-    case ClusterUpdateStatus.ErrorRetrieving:
-      return <ErrorRetrievingMessage cv={cv} />;
-    case ClusterUpdateStatus.Failing:
-      return <FailingMessage cv={cv} />;
-    default:
-      return <UpToDateMessage />;
-  }
 };
 
 export const CurrentVersion: React.FC<CurrentVersionProps> = ({ cv }) => {
@@ -1400,20 +1218,6 @@ export const ClusterSettingsPage: React.FC = () => {
   );
 };
 
-type UpdateStatusProps = {
-  cv: ClusterVersionKind;
-};
-
-type CVStatusMessagePopoverProps = {
-  bodyContent: string;
-  children: React.ReactNode;
-};
-
-type CVStatusMessageProps = {
-  cv: ClusterVersionKind;
-  isFailing?: boolean;
-};
-
 type CurrentChannelProps = {
   cv: K8sResourceKind;
   canUpgrade: boolean;
@@ -1524,10 +1328,6 @@ type ClusterSettingsAlertsProps = {
 type ClusterVersionDetailsTableProps = {
   obj: ClusterVersionKind;
   autoscalers?: K8sResourceKind[];
-};
-
-type ClusterVersionConditionsLinkProps = {
-  cv: ClusterVersionKind;
 };
 
 type ClusterOperatorTabPageProps = {
