+++ frontend/public/locales/en/public.json
@@ -134,16 +134,16 @@
   "Conditions": "Conditions",
   "Operand versions": "Operand versions",
   "Related objects": "Related objects",
-  "Not configured": "Not configured",
+  "View conditions": "View conditions",
   "Invalid cluster version": "Invalid cluster version",
   "Cancel update": "Cancel update",
   "Release not accepted": "Release not accepted",
   "Available updates": "Available updates",
   "Failing": "Failing",
-  "View conditions": "View conditions",
   "Update to {{version}} in progress": "Update to {{version}} in progress",
   "Not retrieving updates": "Not retrieving updates",
   "Up to date": "Up to date",
+  "Not configured": "Not configured",
   "Select a version": "Select a version",
   "Current version": "Current version",
   "Last completed version": "Last completed version",
